requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0

# Unsloth for fast fine-tuning (without problematic extras)
unsloth @ git+https://github.com/unslothai/unsloth.git

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.64.0

# Training utilities
bitsandbytes>=0.41.0
scipy>=1.10.0

# Configuration and logging
python-dotenv>=1.0.0
pyyaml>=6.0.0
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0